[FIX] core: stream Binary fields with non-b64 value
Write some bytes in a `field.Binary(attachment=False)` field. The ORM
doesn't encode the bytes in b64, and the value is stored as a binary
blob in postgres.

Attempt to download the content via /web/content (actually any route
that uses `http.Stream` is affected). It sometimes download something,
sometimes fail with an "Incorrect padding" error.

The `http.Stream` class wrongly assumes that reading a binary/image
field is always going to return the value base64-encoded, thus it always
attemps to decode it. The `b64decode` function silently discard non-b64
characters and only complain if the final thing lacks the b64 `=`
padding (to make the length a multiple of 4).

So when it downloaded something, it downloaded crap.
The thing actually is:

* Binary fields, accept raw bytes, store raw bytes in db.
* Image fields, reject raw bytes, want base64, store base64 in db.

When reading Image fields it is easy: always decode the base64.

For Binary fields it is complicated, because some crazy people encode
their binary fields in base64, and expect automatic base64 decoding
when read. Crazy! So *attempt* to decode the b64 and if the decoding
fail just assume it was raw bytes from the begining.

closes odoo/odoo#253210

closes odoo/odoo#253382

X-original-commit: 37ddfef7e7d2476f74db30819ccc1a2c8c26712d
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>
Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>