[IMP] website: show WebsiteLoader when installing a module from builder
Installing a module from the "New" button (in systray) already displays
the `WebsiteLoader`, but installing from the builder sidebar or "Add
Snippet" dialog did not.

To make the behaviour consistent, we now also show the `WebsiteLoader`
when installing a module from the builder sidebar or "Add Snippet"
dialog.

task-4252688

closes odoo/odoo#193974

Related: odoo/enterprise#77485
Signed-off-by: Francois Georis (fge) <fge@odoo.com>