[FIX] resource: compute date_to from date_from on a resource time off
When creating a public Time off for a Working Schedule, the end date
will automatically set to a moment earlier than the set start date
if the start date hour is set to midnight sharp.

1. Go to Employees.
2. Click on Configuration tab > Working Schedules.
3. Select a Working Schedule.
4. A smart button Public Time Off should appear. Click on it.
5. Click New and select a Start Date with a random date but
00:00:00 as hours:minutes:seconds.

The End date should automatically set 23 hours 59 minutes and
59 seconds later.

A Validation error occurs and the end date is set 1 second before the
start date.

When setting automatically the end date of a leave, it is set to
23:59:59. With our timezone, if we select midnight, this time will
be converted to 23:00:00, the day before:
https://github.com/odoo/odoo/blob/e85f1a182ee60a25905369c84ced05480c5a3360/addons/resource/models/resource_calendar_leaves.py#L69-L70
Thus, the resulting end date will be set to the day before, at
23:59:59.
Additionally, we have to take into account the timezone of the user
to set a consistent end date, and not the generic utc one.

__
opw-5187977

closes odoo/odoo#237362

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>