[IMP] htm_builder, website: handle errors on outdated snippet
Since [1], the warning to inform the user that a snippet is outdated has been removed.

After this commit, if a builder action fail on outdated snippets (mismatched vcss/vxml/vjs
versions), we show a warning notification instead of propagating the error. The notification
informs the user that an error occurred on an outdated snippet, and prompts them to drag
a new version of the snippet.

This is achieved by wrapping try-catch blocks around all builder action lifecycle methods
(apply, clean, load, getValue, isApplied) and implementing an error handling method that
detects if the snippet is outdated.

task-4297808

[1]: https://github.com/odoo/odoo/pull/224382

closes odoo/odoo#230816

Signed-off-by: Francois Georis (fge) <fge@odoo.com>