[FIX] purchase: fetch discount from vendor in catalog
backport of #224231 and #217285

In version 17 and 18 when adding a product from the product catalog
the unit price is set to price_discounted and no disconunt is set
on the purchase order line. The behaviour is not the same when you
add a product in the product selector.

This commit aims to add the discount and price from the catalog
instead, syncronizing the behaviour of catalog and selector. This
issue has already been addressed from 18.2 up, so this is a simple
backport of the fix applied there

opw-5266131

Part-of: odoo/odoo#238331
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>